requirements.txt
openai
python-dotenv
beautifulsoup4 
lxml
playwright